src/coordinator/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sse-starlette==2.1.3
python-multipart==0.0.20
pydantic-settings==2.6.1
openai==1.59.6
anthropic==0.42.0
google-generativeai==0.8.3
pdfplumber==0.11.7
